streaming/requirements.txt
kafka-python==2.2.11
redis==6.2.0
pandas==2.2.2
boto3==1.34.69
pyarrow==20.0.0
s3fs==2024.5.0